Use MinimapType for client and server

+++ src/server.cpp
@@ -2686,7 +2686,7 @@ void Server::SendMinimapModes(session_t peer_id,
 	pkt << (u16)modes.size() << (u16)wanted_mode;
 
 	for (auto &mode : modes)
-		pkt << mode.type << mode.label << mode.size << mode.texture << mode.scale;
+		pkt << (u16)mode.type << mode.label << mode.size << mode.texture << mode.scale;
 
 	Send(&pkt);
 }
